% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiS.R
\name{CondiS}
\alias{CondiS}
\title{CondiS Function}
\usage{
CondiS(time, status, tmax)
}
\arguments{
\item{time}{The follow up time for right-censored data.}

\item{status}{The censoring indicator, normally 0=right censored, 1=event at time.}
}
\description{
This function allows you to impute survival time.
}
\examples{
CondiS(survival_dat)
}
